consolidated the two scripts
